requirements.txt
selenium
scrapy
beautifulsoup4
requests
python-dotenv